ParticipantBox -> ParticipantView, split to UI and core component
