--- packages/react-sdk/src/core/components/ParticipantBox/index.ts
@@ -1,1 +1,0 @@
-export * from './ParticipantBox';
