+++ packages/react-sdk/src/core/components/ParticipantView/index.ts
@@ -1,0 +1,2 @@
+export * from './ParticipantView';
+export * from './DefaultParticipantViewUI';
