+++ packages/react-sdk/src/core/components/index.ts
@@ -1,5 +1,5 @@
 export * from './Audio';
-export * from './ParticipantBox';
+export * from './ParticipantView';
 
 export { Video } from './Video';
 export type { VideoProps } from './Video';
