+++ packages/react-sdk/src/core/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useAudioPublisher';
 export * from './useVideoPublisher';
+export * from './useTrackElementVisibility';
