+++ packages/styling/index.scss
@@ -26,7 +26,7 @@
 @import 'src/StreamCall';
 @import 'src/Tooltip';
 @import 'src/Video';
-@import 'src/ParticipantBox';
+@import 'src/ParticipantView';
 @import 'src/CallLayout';
 @import 'src/Icon';
