--- packages/styling/src/ParticipantBox/index.scss
@@ -1,1 +1,0 @@
-@import './ParticipantBox-layout';
