+++ packages/styling/src/ParticipantView/ParticipantView-layout.scss
@@ -1,22 +1,30 @@
-$scope-name: 'str-video__participant';
+$scope-name: 'str-video__participant-view';
 
-.str-video__participant,
-.str-video__participant-angular-host {
+.str-video__participant-view,
+.str-video__participant-view--angular-host {
   position: relative;
   width: 100%;
   max-width: 1280px;
-  min-height: 0;
   max-height: 100%;
   aspect-ratio: 4/3;
 
-  .str-video__participant_details,
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: var(--str-video__border-radius-sm);
+
+  &--speaking {
+    outline: 2px solid var(--str-video__primary-color);
+  }
+
+  .str-video__participant-details,
   .str-video__call-controls__button {
     transition: opacity 200ms ease-out;
     opacity: 0.3;
   }
 
   &:hover {
-    .str-video__participant_details,
+    .str-video__participant-details,
     .str-video__call-controls__button {
       opacity: 1;
     }
@@ -30,38 +38,16 @@ $scope-name: 'str-video__participant';
     padding: 0.3rem;
   }
 
-  &--speaking {
-    video,
-    .str-video__participant-placeholder {
-      outline: 2px solid var(--str-video__primary-color);
-    }
-  }
-
-  .str-video__video-container {
-    position: relative;
-    height: 100%;
-    min-height: 0;
-    max-height: 100%;
-    display: flex;
-    flex-direction: column;
-
-    .str-video__remote-video {
-      border-radius: var(--str-video__border-radius2);
-      width: 100%;
-      object-fit: cover;
-      background: var(--str-video__text-color3);
-    }
-  }
-
-  .str-video__participant_details {
+  .str-video__participant-details {
     position: absolute;
     left: 0.875rem;
     bottom: 0.875rem;
     display: flex;
     align-items: center;
     gap: 0.3125rem;
+    z-index: 1;
 
-    .str-video__participant_name {
+    .str-video__participant-details__name {
       border-radius: var(--str-video__border-radius1);
       background-color: var(--str-video__background-color4);
       display: flex;
@@ -81,13 +67,13 @@ $scope-name: 'str-video__participant';
         background-color: var(--str-video__text-color1);
       }
 
-      .str-video__participant_name--audio-muted {
+      .str-video__participant-details__name--audio-muted {
         @include icon();
         mask-image: var(--str-video__icon--mic-off);
         -webkit-mask-image: var(--str-video__icon--mic-off);
       }
 
-      .str-video__participant_name--dominant_speaker {
+      .str-video__participant-details__name--dominant_speaker {
         @include icon();
         mask-image: var(--str-video__icon--dominant-speaker);
         -webkit-mask-image: var(--str-video__icon--dominant-speaker);
@@ -100,13 +86,13 @@ $scope-name: 'str-video__participant';
         -webkit-mask-position: center;
       }
 
-      .str-video__participant_name--video-muted {
+      .str-video__participant-details__name--video-muted {
         @include icon();
         mask-image: var(--str-video__icon--camera-off);
         -webkit-mask-image: var(--str-video__icon--camera-off);
       }
 
-      .str-video__participant_name--pinned {
+      .str-video__participant-details__name--pinned {
         @include icon();
         width: 0.8rem;
         -webkit-mask-size: 0.8rem;
@@ -120,7 +106,7 @@ $scope-name: 'str-video__participant';
         -webkit-mask-image: var(--str-video__icon--pin);
       }
 
-      .str-video__participant_name--network-stats {
+      .str-video__participant-details__name--network-stats {
         @include icon();
         background-image: var(--str-video__icon--network-quality);
         cursor: pointer;
@@ -128,7 +114,7 @@ $scope-name: 'str-video__participant';
       }
     }
 
-    .str-video__participant__connection-quality {
+    .str-video__participant-details__connection-quality {
       width: 1.5rem;
       height: 1.5rem;
       display: block;
