+++ packages/styling/src/ParticipantView/index.scss
@@ -1,0 +1,1 @@
+@import './ParticipantView-layout';
